back-auth/requirements.txt
fastapi==0.110.0
authlib==1.3.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
cassandra-driver==3.29.1
aiosmtplib==1.1.6
pydantic-settings==2.1.0
email-validator==2.2.0
httpx==0.27.0
psycopg2-binary==2.9.9
uvicorn==0.29.0
